requirements_dev.txt
atomicwrites==1.3.0
attrs==19.3.0
coverage==4.5.4
ImageHash==4.0
importlib-metadata==0.23
more-itertools==7.2.0
numpy==1.17.4
packaging==19.2
Pillow==6.2.1
pluggy==0.13.0
py==1.8.0
pyparsing==2.4.5
pytest==5.2.2
pytest-cov==2.8.1
PyWavelets==1.1.1
scipy==1.3.3
six==1.13.0
wcwidth==0.1.7
zipp==0.6.0
